rename and start cleaning up tools
