+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=61.0"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "kimera_vio_evaluation"
+name = "kimera_eval"
 version = "0.2.0"
 authors = [
   { name = "Antoni Rosinol", email="arosinol@mit.edu" },
