+++ src/kimera_eval/core/trajectory_metrics.py
