+++ src/kimera_eval/core/trajectory_tools.py
