+++ src/kimera_eval/experiments/example_euroc.yaml
