+++ src/kimera_eval/experiments/full_euroc.yaml
