+++ src/kimera_eval/tools/__init__.py
@@ -1,0 +1,1 @@
+"""Tools and utilities for evaluation."""
